Improvement: Reorganized Location of Pronoun Data Class for Ease of Access
