+++ MekHQ/src/mekhq/campaign/randomEvents/personalities/enums/Greed.java
@@ -38,7 +38,7 @@ import static mekhq.utilities.MHQInternationalization.getFormattedTextAt;
 
 import megamek.common.enums.Gender;
 import megamek.logging.MMLogger;
-import mekhq.campaign.randomEvents.personalities.PersonalityController.PronounData;
+import mekhq.campaign.personnel.PronounData;
 
 /**
  * Represents various levels and traits of greed in a personality.
